+++ packages/commerce-sdk-react/src/hooks/ShopperGiftCertificates/queryKeyHelpers.ts
@@ -6,7 +6,8 @@
  */
 import type {ShopperGiftCertificates} from 'commerce-sdk-isomorphic'
 import {Argument, ExcludeTail} from '../types'
-import {pick} from '../utils'
+import {getCustomKeys, pick} from '../utils'
+import paramKeysMap from './paramKeys'
 
 // We must use a client with no parameters in order to have required/optional match the API spec
 type Client = ShopperGiftCertificates<{shortCode: string}>
@@ -24,11 +25,6 @@ export type QueryKeys = {
 // This is defined here, rather than `types.ts`, because it relies on `Client` and `QueryKeys`,
 // and making those generic would add too much complexity.
 type QueryKeyHelper<T extends keyof QueryKeys> = {
-    /**
-     * Reduces the given parameters (which may have additional, unknown properties) to an object
-     * containing *only* the properties required for an endpoint.
-     */
-    parameters: (params: Params<T>) => Params<T>
     /** Generates the path component of the query key for an endpoint. */
     path: (params: Params<T>) => ExcludeTail<QueryKeys[T]>
     /** Generates the full query key for an endpoint. */
@@ -36,15 +32,14 @@ type QueryKeyHelper<T extends keyof QueryKeys> = {
 }
 
 export const getGiftCertificate: QueryKeyHelper<'getGiftCertificate'> = {
-    parameters: (params) => pick(params, ['organizationId', 'siteId']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
         params.organizationId,
         '/gift-certificate'
     ],
-    queryKey: (params: Params<'getGiftCertificate'>) => [
-        ...getGiftCertificate.path(params),
-        getGiftCertificate.parameters(params)
-    ]
+    queryKey: (params: Params<'getGiftCertificate'>) => {
+        const paramKeys = [...paramKeysMap['getGiftCertificate'], ...getCustomKeys(params)]
+        return [...getGiftCertificate.path(params), pick(params, paramKeys)]
+    }
 }