+++ packages/commerce-sdk-react/src/hooks/ShopperSearch/queryKeyHelpers.ts
@@ -6,8 +6,8 @@
  */
 import type {ShopperSearch} from 'commerce-sdk-isomorphic'
 import {Argument, ExcludeTail} from '../types'
-import {pick} from '../utils'
-
+import {getCustomKeys, pick} from '../utils'
+import paramKeysMap from './paramKeys'
 // We must use a client with no parameters in order to have required/optional match the API spec
 type Client = ShopperSearch<{shortCode: string}>
 type Params<T extends keyof QueryKeys> = Partial<Argument<Client[T]>['parameters']>
@@ -31,11 +31,6 @@ export type QueryKeys = {
 // This is defined here, rather than `types.ts`, because it relies on `Client` and `QueryKeys`,
 // and making those generic would add too much complexity.
 type QueryKeyHelper<T extends keyof QueryKeys> = {
-    /**
-     * Reduces the given parameters (which may have additional, unknown properties) to an object
-     * containing *only* the properties required for an endpoint.
-     */
-    parameters: (params: Params<T>) => Params<T>
     /** Generates the path component of the query key for an endpoint. */
     path: (params: Params<T>) => ExcludeTail<QueryKeys[T]>
     /** Generates the full query key for an endpoint. */
@@ -43,42 +38,29 @@ type QueryKeyHelper<T extends keyof QueryKeys> = {
 }
 
 export const productSearch: QueryKeyHelper<'productSearch'> = {
-    parameters: (params) =>
-        pick(params, [
-            'organizationId',
-            'siteId',
-            'q',
-            'refine',
-            'sort',
-            'currency',
-            'locale',
-            'expand',
-            'offset',
-            'limit'
-        ]),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
         params.organizationId,
         '/product-search'
     ],
-    queryKey: (params: Params<'productSearch'>) => [
-        ...productSearch.path(params),
-        productSearch.parameters(params)
-    ]
+    queryKey: (params: Params<'productSearch'>) => {
+        const paramKeys = [...paramKeysMap['productSearch'], ...getCustomKeys(params)]
+
+        return [...productSearch.path(params), pick(params, paramKeys)]
+    }
 }
 
 export const getSearchSuggestions: QueryKeyHelper<'getSearchSuggestions'> = {
-    parameters: (params) =>
-        pick(params, ['organizationId', 'siteId', 'q', 'limit', 'currency', 'locale']),
     path: (params) => [
         '/commerce-sdk-react',
         '/organizations/',
         params.organizationId,
         '/search-suggestions'
     ],
-    queryKey: (params: Params<'getSearchSuggestions'>) => [
-        ...getSearchSuggestions.path(params),
-        getSearchSuggestions.parameters(params)
-    ]
+    queryKey: (params: Params<'getSearchSuggestions'>) => {
+        const paramKeys = [...paramKeysMap['getSearchSuggestions'], ...getCustomKeys(params)]
+
+        return [...getSearchSuggestions.path(params), pick(params, paramKeys)]
+    }
 }