+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v1.4.0-dev (Jan 22, 2024)
+- Fix invalid query params warnings and allow custom query [#1655](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1655)
+
 ## v1.3.0 (Jan 19, 2024)
 
 - Add support for node 20 [#1612](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1612)